Features completed except for the bonus challenge for excuding duplicates of position. Next: styling

+++ app/nfl-roster/playerController.js
@@ -3,7 +3,10 @@ function PlayerController() {
     var playerService = new PlayerService(ready);
     var currentFilter = '';
     var currentSelection = '';
+
+    //---------------------------------------------------------
     // PRIVATE
+    //---------------------------------------------------------
 
     function ready() {
         loading = false; //stop the spinner
@@ -17,7 +20,10 @@ function PlayerController() {
     // });
     // console.log(playerService.getPlayersByTeam("SF"));
 
+    //---------------------------------------------------------
+    // POPULATE DROPDOWN MENUS
     // EACH DROP DOWN MENU DYNAMICALLY GENERATES ITS SELECTIONS
+    //---------------------------------------------------------
     function renderLastnamesDropdown() {
         var elem = document.getElementById('lastnames-list');
         var template = '';
@@ -71,6 +77,17 @@ function PlayerController() {
     }
     renderTeamsDropdown()
 
+    //---------------------------------------------------------
+    // UPDATE / REFRESH FUNCTIONS FOR ROSTER AND TEAM
+    // updateRosterDisplay() TAKES A POST SEARCH FILTERED
+    // LIST
+    // refreshTeamDisplay() TAKES THE PERSISTENT currentFilter
+    // and currentSelection variables 
+    // (Updated each time you change searches) IN ORDER TO
+    // RESORT THE ROSTER AND THEN CALL updateRosterDisplay()
+    // AGAIN AFTER SPLICING OUT A SELECTION.
+    //---------------------------------------------------------
+
     function updateRosterDisplay(filteredPlayerList) {
         var elem = document.getElementById('player-roster')
         var template = ''
@@ -79,11 +96,11 @@ function PlayerController() {
                 var filteredPlayer = filteredPlayerList[key];
                 template += `
                      <div class="player-card">
-                          <img src="${filteredPlayer.photo}" alt="Player image placeholder" id="player-image">
+                          <img src="${filteredPlayer.photo}" alt="Player image placeholder" class="player-image">
                           <h4 class="player-name">${filteredPlayer.fullname}</h4>
                           <h6 class="player-position">${filteredPlayer.position}</h6>
                           <h6 class="player-team">${filteredPlayer.pro_team}</h6>
-                          <button class="addButton" id="${filteredPlayer.id}" onclick="playerController.addToRoster(${filteredPlayer.id})">Add to team</button>
+                          <button class="addButton" id="${filteredPlayer.id}" onclick="playerController.addToTeam(${filteredPlayer.id})">Add to team</button>
                      </div>
                 `
             }
@@ -92,7 +109,7 @@ function PlayerController() {
 
     }
 
-    function refreshTeamDisplay(currentFilter, currentSelection) {
+    function refreshRosterDisplay(currentFilter, currentSelection) {
         switch (currentFilter) {
             case 'lastname':
                 var tempPlayerList = playerService.getPlayersByLastName(currentSelection);
@@ -112,6 +129,27 @@ function PlayerController() {
         }
     }
 
+    function updateTeamDisplay() {
+        var teamList = playerService.getCurrentTeam();
+        var elem = document.getElementById('my-team')
+        var template = ''
+        for (var key in teamList) {
+            if (teamList.hasOwnProperty(key)) {
+                var teamPlayer = teamList[key];
+                template += `
+                     <div class="player-card">
+                          <img src="${teamPlayer.photo}" alt="Player image placeholder" class="player-image">
+                          <h4 class="player-name">${teamPlayer.fullname}</h4>
+                          <h6 class="player-position">${teamPlayer.position}</h6>
+                          <h6 class="player-team">${teamPlayer.pro_team}</h6>
+                          <button class="addButton" id="${teamPlayer.id}" onclick="playerController.removeFromTeam(${teamPlayer.id})">Remove from team</button>
+                     </div>
+                `
+            }
+        }
+        elem.innerHTML = template;
+
+    }
 
     // function renderSortBy(sortBy) { // STALLED ATTEMPT TO RENDER SUB-SEARCH MENU, MAY REVISIT LATER
     //     var elem = document.getElementById('sortByList')
@@ -119,19 +157,28 @@ function PlayerController() {
     // var sortedPlayers = playerService.
     // }
 
-    // PUBLIC
 
-    // GET PLAYERS BY SEARCH TYPE (Not sure what I was thinking when I wrote these public.
-    // Not in use. leaving here in case it comes back to me.)
+    //---------------------------------------------------------
+    // PUBLIC
+    //---------------------------------------------------------
+
+    //---------------------------------------------------------
+    //  renderByX FUNCTIONS RECIEVE THE ENTIRE FORM THAT YOU 
+    // SEARCHED WITH (ONCHANGE, USING this.form) THEY THEN
+    // ASSIGN THE INDEXED CHOICE AND THEN THE VALUE IN THAT 
+    // INDEX TO VARIABLES, WHICH IS THEN USED TO GET A
+    // SORTED PLAYER LIST. THE FUNCTION THEN SETS PERSISTENT
+    // FILTER TYPE AND SELECTION VALUES BEFORE PASSING THE
+    // SORTED ARRAY TO updateRosterDisplay() TO DRAW THE
+    // SORTED DATA TO THE ROSTER
+    //---------------------------------------------------------
 
     this.renderByLastname = function renderByLastname(lastNameForm) {
         // console.log("Last name form: ", lastNameForm)
         var choiceIndex = lastNameForm.searchByLastnameList.selectedIndex;
         var choiceValue = lastNameForm.searchByLastnameList.options[choiceIndex].value;
         var sortedChoices = playerService.getPlayersByLastName(choiceValue)
-        console.log(sortedChoices)
-        // START HERE, WRITE A FUNCTION THAT IS CALLED IN EACH OF THESE 3
-        // THAT WILL UPDATE THE SCREEN WITH THE SORTED ROSTER
+        console.log("Current Roster Sort (By last name): ", sortedChoices)
         currentFilter = 'lastname';
         currentSelection = choiceValue;
         // console.log(currentFilter, currentSelection)
@@ -139,32 +186,53 @@ function PlayerController() {
     }
 
     this.renderByPosition = function renderByPosition(position) {
-        console.log("Position form: ", position)
+        // console.log("Position form: ", position)
         var choiceIndex = position.searchByPositionList.selectedIndex;
         var choiceValue = position.searchByPositionList.options[choiceIndex].value;
         var sortedChoices = playerService.getPlayersByPosition(choiceValue)
-        console.log(sortedChoices)
+        console.log("Current Roster Sort (By position): ", sortedChoices)
         currentFilter = 'position';
         currentSelection = choiceValue;
         updateRosterDisplay(sortedChoices);
     }
 
     this.renderByTeam = function renderByTeam(team) {
-        console.log("Team form: ", team)
+        // console.log("Team form: ", team)
         var choiceIndex = team.searchByTeamList.selectedIndex;
         var choiceValue = team.searchByTeamList.options[choiceIndex].value;
         var sortedChoices = playerService.getPlayersByTeam(choiceValue)
-        console.log(sortedChoices)
+        console.log("Current Roster Sort (By team): ", sortedChoices)
         currentFilter = 'team';
         currentSelection = choiceValue;
         updateRosterDisplay(sortedChoices);
     }
 
-    this.addToRoster = function addToRoster(selectedPlayerId) {
-        playerService.addToRoster(selectedPlayerId);
-        playerService.removeFromRoster(selectedPlayerId);
-        refreshTeamDisplay(currentFilter, currentSelection);
+    //---------------------------------------------------------
+    // ADD TO / REMOVE FROM ROSTER / TEAM FUNCTIONS
+    //---------------------------------------------------------
+
+    this.addToTeam = function addToTeam(selectedPlayerId) {
+        if (playerService.getCurrentTeam().length < 16) {
+            playerService.addToTeam(selectedPlayerId);
+            playerService.removeFromRoster(selectedPlayerId);
+            refreshRosterDisplay(currentFilter, currentSelection);
+            updateTeamDisplay();
+        }
     }
+
+    this.removeFromTeam = function removeFromTeam(selectedPlayerId) {
+        playerService.addToRoster(selectedPlayerId);
+        playerService.removeFromTeam(selectedPlayerId);
+        refreshRosterDisplay(currentFilter, currentSelection);
+        updateTeamDisplay();
+    }
+
+
+
+
+    //---------------------------------------------------------
+    // UNNEEDED / COME BACK TO LATER
+    //---------------------------------------------------------
 
     // *** ATTEMPTED TO COMBINE SORT FUNCTION INTO ONE, FELL BACK TO SEPARATE, MAY REVISIT LATER
     // this.getSortedPlayers = function getSortedPlayers(field, sortBy) {
